Added a new "utilities" file
The new utilities file will hold simple functions to be used in several places throughout MADRID

Added a new "stringlist_to_list" function. This function takes a "string list" ("['mat', 'xml', 'json']") and converts it into a list (["mat", "xml", "json"]). This is to be used to deprecate the "string list" behavior