Add GitHub action to publish Jupyter notebook
A new GitHub Actions workflow has been added for publishing the COMO notebook to GitHub Pages. The workflow triggers on push events to the 'feature/jupyter-pages' branch. It runs on the latest Ubuntu and uses Miniconda to manage Python and other dependencies. The environment is set up to automatically update conda and use Python 3.10. It proceeds to install the necessary packages as defined in the 'environment.yaml' file.