Added conda initialization to GitHub workflow
The installation script in the GitHub workflow file 'publish_notebook.yml' has been modified to first initialize conda and source the bashrc file, before proceeding with the activation of the base conda environment and installation of dependencies. This change ensures that the conda commands are recognized and carried out properly during the GitHub Action's runtime, preventing potential errors if conda is not in the PATH by default.