Rename job and adjust conda channels in GitHub workflow
The changes in `.github/workflows/publish_notebook.yml` include renaming the 'deploy-book' job to 'deploy-como' and altering the conda channels. The new conda channels list, which includes 'bioconda', 'r', 'gurobi', along with the default ones will give us access to more packages needed for our tasks. This improves the flexibility and robustness of the workflow setup.