Update Micromamba installation method in workflows
The Micromamba installation method in the GitHub workflow 'publish_notebook.yml' was changed. Instead of using 'wget' to download the install.sh script, we now directly download the latest linux-64 version of Micromamba using a HTTP request. This will ensure that the most recent version of Micromamba is always used in the workflow.