Update package listing command and integrate Jupyter-Book with Github Pages
This commit switches the package listing command from mamba list to micromamba list for compatibility purposes. Functionality for Jupyter-Book is also added. This allows for the built book to be pushed to Github Pages, thus enhancing project visibility and accessibility. It requires an install command and a build command for the jupyter-book, followed by a push command which uses Github Actions to deploy the built book to Github Pages.