Update Github workflow for Jupyter notebook publishing
Changed the Github Workflow for publishing Jupyter notebooks. It now uses Jupyterlite for building the notebook instead of Jupyter Book. This allows for a faster build time as it no longer requires to build a table of contents. The deployment to Github Pages was also separated into a different job, called "deploy-como", for improved modularity and efficiency. It deploys only when "build-como" job has completed successfully.