Simplify paths in Github Actions workflow
The Github Actions workflow script was simplified by removing explicit references to the github workspace from the JupyterLite build command and artifact upload path. The new changes made the scripts easier to read and modify. Also, the "List build items" step has been removed as it was found to be redundant.