Fix broken Jupyterlite installation in workflow
Replaced the default system path for micromamba with an absolute path in the Github Actions workflow. The change was made due to inconsistency in environment settings as the system path was not always correctly set. Using the absolute path for micromamba ensures the installation of Jupyterlite will be executed correctly regardless of the environment setup.