Update file paths in GitHub workflow
The paths for the environment file and the pip install requirements in the GitHub workflow file (unit_tests.yml) were incorrect. These have been updated to main/src/tests from main/py/tests to ensure that the unit tests are able to run correctly in the GitHub environment. This change fixes the Github actions workflow for automated testing.