Update installation instructions in README.md
Switched the project from a Docker-based setup to a Conda-based setup in README.md, making it easier for users to set up and run the project. Added instructions for installing Conda and Mamba, cloning the repository, creating and activating a new Conda environment, and starting the notebook server. This shift is expected to speed up the installation process and provide a smoother user experience, also it's more transparent for debugging when issues occur. Additionally, re-arranged the sections related to documentation and resources for better readability.